Refactor test container
This hides the "Multiple definitions exist for class 'TestKernel'"
inspection message in PHPStorm.
