+++ phpstan-baseline.neon
@@ -41,16 +41,6 @@ parameters:
 			path: src/Database/RunCommand.php
 
 		-
-			message: "#^Method Pimcore\\\\Kernel\\:\\:configureContainer\\(\\) invoked with 1 parameter, 3 required\\.$#"
-			count: 1
-			path: src/Kernel/TestKernel.php
-
-		-
-			message: "#^Parameter \\#1 \\$bundle of method Pimcore\\\\HttpKernel\\\\BundleCollection\\\\BundleCollection\\:\\:addBundle\\(\\) expects string\\|Symfony\\\\Component\\\\HttpKernel\\\\Bundle\\\\BundleInterface, Pimcore\\\\Bundle\\\\AdminBundle\\\\PimcoreAdminBundle given\\.$#"
-			count: 1
-			path: src/Kernel/TestKernel.php
-
-		-
 			message: "#^Parameter \\#1 \\$name of static method Neusta\\\\Pimcore\\\\TestingFramework\\\\Pimcore\\\\BootstrapPimcore\\:\\:setEnv\\(\\) expects string, int\\|string given\\.$#"
 			count: 1
 			path: src/Pimcore/BootstrapPimcore.php
