+++ phpstan.neon
@@ -9,8 +9,8 @@ parameters:
 		- src
 
 	excludePaths:
-		# There are two errors, that were not able to add to the baseline
-		- src/Kernel/TestKernel.php
+		# There are two errors, that were not able to be added to the baseline
+		- src/Kernel/CompatibilityKernel.php
 
 	bootstrapFiles:
 		- vendor/pimcore/pimcore/stubs/dynamic-constants.php
